Move services related to the virtual file system to a separate registry
