+++ subprojects/core/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.PluginServiceRegistry
@@ -1,2 +1,3 @@
 org.gradle.caching.internal.BuildCacheServices
 org.gradle.internal.service.scopes.ExecutionServices
+org.gradle.internal.service.scopes.VirtualFileSystemServices
